refactor: useSiteMetadata hook
