+++ tsconfig.json
@@ -37,7 +37,6 @@
       "@module/*": ["src/module/*"],
       "@pages/*": ["src/pages/*"],
       "@scss/*": ["src/scss/*"],
-      "@services/*": ["src/services/*"],
       "@store/*": ["src/store/*"],
       "@hooks/*": ["src/hooks/*"],
       "@types": ["src/types"],
